Move RecoverableHeat to utils test
